Audio and menu inclusion
Created game.gd to set exit button and get background music working
Added control node and child to have exit button on main game

Created mainmenu scene and script

Created options menu scene and script
added entries to shared to facilitate consistant volume throughout a single game instance, not sure how to save that data to a file for next use yet

Created 2 additional volume busses: sfx and volume

Added following audiostreams:
lineclear to board (child node of game.tscn)
rotationsound, movesound, clunksound all to tetromino.tscn

Changed _input(event) in tetromino:
instead of example:
		move(Vector2.RIGHT)

it's now
		if(move(Vector2.RIGHT)):
		move_sound.play()

as it still triggers the move function to make the appropriate positional adjustments but only makes a noise if the move is "legal" and the piece is actually moved. If I just added it as another line on the original way it played constantly. If I added the sound to the move function it played every time it ticked down with the timer and also every time the player pressed a button which was annoying after like 15 seconds of testing.

**Wanted to do that method for all sounds but it didn't work with the rotate or the hard drop as neither have a userful (non-void) return type

Fixed tetriminos appearing above the top row of grid border tiles by changing it's z-value in the TileMapLayer of GameGrid (gridv2.tscn)

Created AudioManager.gd and added to autoload so clicks kept playing between scenes when pressed, watched a youtube tutorial for that one so not entirely sure what is going on: